Factor-out testcases into separate files
